+++ lib/api/src/grpc/qdrant.rs
@@ -989,6 +989,9 @@ pub struct TurboQuantization {
     /// Number of Hadamard rotations (only used when rotation is Hadamard)
     #[prost(uint32, optional, tag = "5")]
     pub hadamard_chunk: ::core::option::Option<u32>,
+    /// Enable turbo quant plus mode
+    #[prost(bool, optional, tag = "6")]
+    pub plus: ::core::option::Option<bool>,
 }
 #[derive(validator::Validate)]
 #[derive(serde::Serialize)]
@@ -2488,7 +2491,7 @@ impl TurboQuantCorrection {
 pub enum TurboQuantRotation {
     TqrHadamard = 0,
     TqrNoRotation = 1,
-    TqrRandom = 2,
+    TqrRotationMatrix = 2,
 }
 impl TurboQuantRotation {
     /// String value of the enum field names used in the ProtoBuf definition.
@@ -2499,7 +2502,7 @@ impl TurboQuantRotation {
         match self {
             TurboQuantRotation::TqrHadamard => "TQR_Hadamard",
             TurboQuantRotation::TqrNoRotation => "TQR_NoRotation",
-            TurboQuantRotation::TqrRandom => "TQR_Random",
+            TurboQuantRotation::TqrRotationMatrix => "TQR_RotationMatrix",
         }
     }
     /// Creates an enum from field names used in the ProtoBuf definition.
@@ -2507,7 +2510,7 @@ impl TurboQuantRotation {
         match value {
             "TQR_Hadamard" => Some(Self::TqrHadamard),
             "TQR_NoRotation" => Some(Self::TqrNoRotation),
-            "TQR_Random" => Some(Self::TqrRandom),
+            "TQR_RotationMatrix" => Some(Self::TqrRotationMatrix),
             _ => None,
         }
     }
