+++ lib/quantization/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod coordinate_analysis;
 pub mod encoded_storage;
 pub mod encoded_vectors;
 pub mod encoded_vectors_binary;
